Removed helper_yocleanup.fish
Its functionality is now in helper_cleanup.fish.
Both functions were doing the same thing, but the yocleanup version
started with replacing 127.0.0.1 with 0.0.0.0 and then cleaning
up the file.
Now all downloaded files go through this even if they don't contain 127.0.0.1.
This will take a little longer but it simplifies the whole setup.
If more lists are added in the future we don't have to worry if it is 127.0.0.1 or 0.0.0.0 based.