Add test for logged in as root on local machine
Why:

* Although it is unlikely on a Mac to have a root user account it's
possible. So lets check for that and display root in reversed red before
the pathname. If we are a normal user (admin or not) we show no
username.
Also if we ever jump to Linux (where some distro's do have a root user)
we will be prepared.

This change addresses the need by:

* Me