Fix prompt with simple chars (fish issue #5501)
Why:

* A permanent fix for commit b74314c

The BLACK RIGHT POINTING TRIANGLE that I used as prompt indicator is a
combining Unicode character, which means it uses 2 char codes to produce
1 character.

This means I have to set non-default settings in Terminal and/or iTerm
and an universal variable in fish to fix fish issue #5501 (polluting
history/prompt with f's and i's). Doing so introduced problems in vim's
statusline.

So I removed the triangle, and replaced it with simple 1 char-wide
code-points characters. It makes for a simpler setup overall.

This change addresses the need by:

* Me