Make format selection more robust
Why:

* Not specifying the exact bitrate but searching for the best quality
makes this more robust. They can change the bitrates anytime they want,
we will just look for the last entry in the list that has a specific
format specifier, and select that. This should keep working as long as
they keep the list in the order it is now, i.e. from lowest to highest
quality.

This change addresses the need by:

* Me