Fix prompt: show user@host when local root & ssh
Why:

* sudo -s is the problem: there is no distinction in environment if we
do a sudo -s locally or over a ssh connection. So when I did a sudo -s
over an ssh connection the prompt only showed 'root' in reversed red,
giving the impression that we are on the local machine. Now we show the
user@host whether we are on a local machine or on a ssh connection.

It is a bit ugly, as I like my prompt to be as minimal as possible. So I
will try to find a way to only show 'root' if we are on a local machine
and 'user@host' if we are on a ssh connection. For the moment this will
have to do.

This change addresses the need by:

* Me