Replace grep with ripgrep (rg) & accept flags args
Why:

* ripgrep is a bit smarter when matches are close together. It won't
repeat lines than have been shown earlier. If the match in the range
given (7 lines) occurs multiple times it starts to recalculate the range
from the last occurrence, and continues on. So I don't have to read the
same line several times in a row.

Also made passing in flags (like -c or --color) possible.

This change addresses the need by:

* Me